refactor: remove unused reviews components, move GetReviewsButton
